Splitting of cf-key
Functionality is moved to libpromises to make it easier for testing.
Only main and getopts will remain in cf-key.c
I added the skeleton of a test unit so we can start populating it
with real tests :-)
