+++ libpromises/Makefile.am
@@ -24,6 +24,7 @@ libpromises_la_SOURCES = \
         cf3.defs.h \
         cf3.extern.h \
         cf3globals.c \
+		cf-key-functions.c cf-key-functions.h \
         cfstream.c cfstream.h \
         chflags.c chflags.h \
         classes.c \
